requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.8
decorator==5.1.1
Django==4.2.3
dnspython==2.3.0
dominate==2.8.0
email-validator==2.0.0.post2
extensions==0.4
filelock==3.12.2
Flask==2.2.5
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
huggingface==0.0.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.202
langchainplus-sdk==0.0.10
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
natural==0.2.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openai-pygenerator==0.4.1
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
protobuf==4.23.3
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
streamlit==1.23.1
tenacity==8.2.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
visitor==0.1.3
watchdog==3.0.0
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
